# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Turnio API
  version: 0.1.0
  description: Contrato vivo entre el backend y el frontend de Turnio. Ver también
    /CONTRATO.md en la raíz del repo para convenciones que este schema no captura
    (auth, formato de errores, etc.).
paths:
...
  /api/agenda/horarios/semana/:
    put:
      operationId: agenda_horarios_semana_update
      description: |-
        Reemplaza el horario **propio** de uno o varios empleados, en una sola transacción.

        El horario propio es la **excepción** al horario del negocio (ver `PUT /api/agenda/horario-negocio/`), para quien no trabaja como el resto: medio tiempo, solo sábados, turno de tarde. Si un empleado no tiene horario propio, hereda el del negocio — que es el caso normal y no requiere llamar a este endpoint.

        `miembros` es una lista porque los de medio tiempo suelen compartir turno. Un solo empleado es el caso `miembros: [id]`.

        Semántica de **reemplazo**: las franjas enviadas pasan a ser el horario propio completo de **cada** empleado señalado. Enviar `franjas: []` le quita el horario propio y lo devuelve a heredar el del negocio; para que alguien no atienda, la palanca es `activo=False` en su membresía, no un horario vacío.

        Todo o nada: si la semana es inválida, o si alguno de los `miembros` no pertenece al negocio, responde `400` y no se toca el horario de nadie.
      tags:
      - agenda
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HorarioSemanal'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/HorarioSemanal'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/HorarioSemanal'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HorarioTrabajo'
          description: ''
...
  /api/caja/ventas/{id}/cobrar/:
    post:
      operationId: caja_ventas_cobrar_create
      description: |-
        Cobra la venta, total o parcialmente. Crea el pago **y** su movimiento de caja en la misma transacción, así que exige caja abierta.

        Un pago mixto son dos llamadas a este endpoint sobre la misma venta, una por método. Cuando lo cobrado alcanza el total, la venta pasa a `pagada` y recién ahí se devengan las comisiones.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this venta.
        required: true
      tags:
      - caja
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cobrar'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Cobrar'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Cobrar'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venta'
          description: ''
...
components:
  schemas:
...
    Cobrar:
      type: object
      description: |-
        Entrada de `POST /api/caja/ventas/{id}/cobrar/`.

        Un pago mixto son **dos llamadas** a este endpoint sobre la misma
        venta, una por método. Se prefirió eso a recibir una lista de pagos
        en un solo request porque cada pago es un hecho independiente que
        puede fallar por su cuenta (la caja se cerró entremedio, el monto
        excede el saldo), y un error parcial dentro de una lista es
        justamente lo que no se quiere en el dominio de dinero.
      properties:
        monto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        metodo_pago:
          $ref: '#/components/schemas/MetodoPagoEnum'
      required:
      - metodo_pago
      - monto
    Devolucion:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        monto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        metodo_pago:
          allOf:
          - $ref: '#/components/schemas/MetodoPagoEnum'
          readOnly: true
        motivo:
          type: string
          readOnly: true
        registrado_por:
          type: integer
          readOnly: true
        registrado_por_nombre:
          type: string
          readOnly: true
        movimiento:
          type: integer
          readOnly: true
        creado_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - creado_en
      - id
      - metodo_pago
      - monto
      - motivo
      - movimiento
      - registrado_por
      - registrado_por_nombre
...
    DiaSemanaEnum:
      enum:
      - 0
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      type: integer
      description: |-
        * `0` - Lunes
        * `1` - Martes
        * `2` - Miércoles
        * `3` - Jueves
        * `4` - Viernes
        * `5` - Sábado
        * `6` - Domingo
...
    FranjaHorario:
      type: object
      description: |-
        Una franja suelta dentro de la semana. Sin `miembro`: lo define el
        contenedor, para no repetirlo en cada elemento de la lista.
      properties:
        dia_semana:
          $ref: '#/components/schemas/DiaSemanaEnum'
        hora_inicio:
          type: string
          format: time
        hora_fin:
          type: string
          format: time
      required:
      - dia_semana
      - hora_fin
      - hora_inicio
...
    HorarioSemanal:
      type: object
      description: |-
        Entrada de `PUT /api/agenda/horarios/semana/`.

        `miembros` es una lista porque lo normal es que todo el equipo comparta
        el mismo horario; un solo empleado es el caso n=1, no una forma aparte.
        Va vacía nunca: sin destinatarios no hay nada que aplicar.

        Semántica de reemplazo: la lista de franjas es el horario completo de
        **cada** empleado señalado, no un agregado. Mandar `franjas: []` los
        deja sin disponibilidad.
      properties:
        miembros:
          type: array
          items:
            type: integer
        franjas:
          type: array
          items:
            $ref: '#/components/schemas/FranjaHorario'
      required:
      - franjas
      - miembros
    HorarioTrabajo:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        miembro:
          type: integer
        dia_semana:
          allOf:
          - $ref: '#/components/schemas/DiaSemanaEnum'
          minimum: -2147483648
          maximum: 2147483647
        hora_inicio:
          type: string
          format: time
        hora_fin:
          type: string
          format: time
      required:
      - dia_semana
      - hora_fin
      - hora_inicio
      - id
      - miembro
...
    MetodoPagoEnum:
      enum:
      - efectivo
      - tarjeta
      - nequi
      - daviplata
      - bre_b
      - otro
      type: string
      description: |-
        * `efectivo` - Efectivo
        * `tarjeta` - Tarjeta
        * `nequi` - Nequi
        * `daviplata` - Daviplata
        * `bre_b` - Bre-B
        * `otro` - Otro
...
    Pago:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        monto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        metodo_pago:
          allOf:
          - $ref: '#/components/schemas/MetodoPagoEnum'
          readOnly: true
        registrado_por:
          type: integer
          readOnly: true
        registrado_por_nombre:
          type: string
          readOnly: true
        movimiento:
          type: integer
          readOnly: true
        creado_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - creado_en
      - id
      - metodo_pago
      - monto
      - movimiento
      - registrado_por
      - registrado_por_nombre
...
    Venta:
      type: object
      description: |-
        Una cuenta: qué se hizo, cuánto vale, cuánto se cobró.

        `estado`, `total` y todo lo relacionado con el cobro son de solo
        lectura: los fija el servidor a través de `cobrar`, `devolver` y
        `anular`. El cliente solo manda el cliente, los items y (si aplica)
        la evidencia.
      properties:
        id:
          type: integer
          readOnly: true
        cita:
          type: integer
          readOnly: true
          nullable: true
        nombre_cliente:
          type: string
          maxLength: 150
        telefono_cliente:
          type: string
          maxLength: 30
        items:
          type: array
          items:
            $ref: '#/components/schemas/VentaItem'
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        total_pagado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        saldo_pendiente:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        estado:
          allOf:
          - $ref: '#/components/schemas/VentaEstadoEnum'
          readOnly: true
        observaciones:
          type: string
        evidencia:
          type: string
          format: uri
          nullable: true
        pagos:
          type: array
          items:
            $ref: '#/components/schemas/Pago'
          readOnly: true
        devoluciones:
          type: array
          items:
            $ref: '#/components/schemas/Devolucion'
          readOnly: true
        creada_por:
          type: integer
          readOnly: true
        creada_por_nombre:
          type: string
          readOnly: true
        anulada_por:
          type: integer
          readOnly: true
          nullable: true
        anulada_en:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        motivo_anulacion:
          type: string
          readOnly: true
        creado_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - anulada_en
      - anulada_por
      - cita
      - creada_por
      - creada_por_nombre
      - creado_en
      - devoluciones
      - estado
      - id
      - items
      - motivo_anulacion
      - nombre_cliente
      - pagos
      - saldo_pendiente
      - total
      - total_pagado
    VentaEstadoEnum:
      enum:
      - pendiente
      - parcial
      - pagada
      - anulada
      type: string
      description: |-
        * `pendiente` - Pendiente de cobro
        * `parcial` - Pagada parcialmente
        * `pagada` - Pagada
        * `anulada` - Anulada
    VentaItem:
      type: object
      description: |-
        Una línea de la venta.

        Al **crear**, se manda `servicio` (y de ahí se copian descripción,
        precio y comisión) o bien `descripcion` + `precio_unitario` a mano,
        para vender algo que todavía no tiene catálogo. `empleado` es siempre
        obligatorio: es la fuente de verdad de la comisión.

        Al **leer**, `descripcion`, `precio_unitario` y `porcentaje_comision`
        son los valores congelados al momento de la venta, no los del
        catálogo de hoy.
      properties:
        id:
          type: integer
          readOnly: true
        servicio:
          type: integer
          nullable: true
        empleado:
          type: integer
        empleado_nombre:
          type: string
          readOnly: true
        descripcion:
          type: string
          maxLength: 200
        precio_unitario:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        cantidad:
          type: integer
          maximum: 2147483647
          minimum: 1
        porcentaje_comision:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          readOnly: true
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - empleado
      - empleado_nombre
      - id
      - porcentaje_comision
      - subtotal
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT